Make using Ins() -arcs safe Ins() arcs are no longer locally greedy, so there is no semantic difference between network inclusion with Ins() or by explicitly embedding. eg. define greedy "a+"; now the following work the same way: Ins(greedy) {abc} greedy {abc}
This also involved cleanup of various parts of the code. While new tests
aren't broken by this, it is not unlikely that such a big change may
reintroduce some bugs or uncover misunderstandings.

This also slows down Ins() arc processing, in some cases by a lot, but
usually not very seriously.