Keep separate RTN stacks for each depth level
Because RTNs can have multiple return sites, it's necessary to keep a
stack of return addresses when returning from them to a lower level in
the rules. In certain combinations this was not being handled correctly,
causing bugs. Now we keep track of a stack of return addresses for each
rule depth level, fixing this.

Also, add placeholder support for weight cutoffs.