hfst-proc: more fixes and tests for serial unicode ranges
a3445f47d7e0561a2a00c4b3033dcd8844f032b5 only fixed the lower part of
Latin Extended A, and messed up the upper part. The bug was a missing
"break" in the every-other-character-is-upper-and-lower pattern of
that block (where a break is introduced by one of the execptions,
e.g. Ÿ or ŉ).

Letters [ıŀŉĸ] still need exceptions, and I haven't even looked at the
other serial ranges …