pmatch: Make INPUT_MARK not special_symbols, so we can refer to it in pmscripts
There's a test in `PmatchAlphabet::is_special`

        return (symbol.find("@PMATCH") == 0 && symbol.at(symbol.size() - 1) == '@')

which makes tags like `@PMATCH_INPUT_MARK@` special, and it seems any "special" symbol can not be referred to as tags in pmatch scripts. However `@PMATCH_INPUT_MARK@` is not special in the same sense as `@PMATCH_ENTRY@` or `@PMATCH_LC_EXIT@` since it occurs in the actual morphologies we include with `@bin"analyser.hfst"`.

That is, if we do

    Define morphology @bin"analyser.hfst" ;  !! contains 1:1 0: Tag 0:@PMATCH_INPUT_MARK@ .:. 0:Tag2
    Define inputmark    [ 0:"@PMATCH_INPUT_MARK@" ];
    Define any [ ? | 0:? | inputmark | flags ];
    Define dotinside      morphology & [ any* {.} any* ] ;

then `dotinside` won't match the analysis (or even if we call it `@PMATCH_JUST_MADE_THIS_UP@` in the hfst and pmscript).

This change makes `@PMATCH_INPUT_MARK@` no longer "special", instead just storing the symbol in `PmatchAlphabet.input_mark_symbol`, letting us match on it in pmscripts.

-----

However, it's sort of an XY problem – I'd prefer to have a way to skip any "invisible" tag sequence (including "PMATCH_INPUT_MARK" or flag diacritics) when intersecting the morphology with a pmatch expression, ie. I'd like to be able to match anything from the morphology that ends in punctuation, regardless of whether it has flags and input-marks after the punctuation and allowing *anything* before the punctuation:

    Define endsInPunct      morphology & [ printableOrUnprintable* Punct unprintable* ] ;